# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StreamPay API",
    "version": "2.0.0",
    "description": "REST API for the StreamPay payment-streaming platform on Stellar.\n\nAll error responses use the canonical envelope:\n```json\n{ \"error\": { \"code\": \"ERROR_CODE\", \"message\": \"Human-readable detail.\", \"request_id\": \"req_...\" } }\n```\n\nv1 stream paths (`/api/streams`) are deprecated and will be removed. Use `/api/v2/streams`."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Obtain a token via POST /api/auth/wallet, then pass it as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code.",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "description": "Human-readable description safe to display to clients.",
                "example": "The requested stream does not exist."
              },
              "request_id": {
                "type": "string",
                "description": "Opaque identifier for this request, useful for log correlation.",
                "example": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
...
      "WalletChallenge": {
        "type": "object",
        "required": [
          "challenge",
          "expires_at"
        ],
        "properties": {
          "challenge": {
            "type": "string",
            "description": "One-time nonce the client must sign with their Stellar private key.",
            "example": "streampay_auth_1716000000000_x7k2m"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Challenge expiry (5 minutes from issuance).",
            "example": "2024-01-15T10:05:00.000Z"
          }
        }
      },
      "WalletToken": {
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque bearer token. Pass as `Authorization: Bearer <token>`.",
            "example": "tok_R0FCQzEyMzQ1Njc4OTA"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Token expiry (24 hours from issuance).",
            "example": "2024-01-16T10:00:00.000Z"
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad request \u2014 invalid input.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "BAD_REQUEST",
                "message": "Request body must include 'recipient' and 'rate'.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized \u2014 missing or invalid bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Bearer token required.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "FORBIDDEN",
                "message": "This endpoint is not available in production.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "INTERNAL_SERVER_ERROR",
                "message": "An unexpected error occurred.",
                "request_id": "req_01HZ9ABCDEF"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/auth/wallet": {
      "get": {
        "operationId": "getWalletChallenge",
        "summary": "Issue a wallet authentication challenge",
        "description": "Returns a one-time nonce the client must sign with their Stellar private key, then submit to `POST /api/auth/wallet`.",
        "security": [],
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "description": "Stellar public key (G\u2026, 56 characters).",
            "schema": {
              "type": "string",
              "pattern": "^G[A-Z2-7]{55}$",
              "example": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletChallenge"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "429": {
            "description": "IP rate limit exceeded for challenge issuance (20 req/min per IP).",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the client may retry.",
                "schema": {
                  "type": "integer"
                }
              },
              "x-request-id": {
                "description": "Correlation ID for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "error": {
                    "code": "rate_limit_exceeded",
                    "message": "Too many requests. Please try again later.",
                    "request_id": "req_01HZ9ABCDEF"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "verifyWalletSignature",
        "summary": "Verify a signed challenge and issue a bearer token",
        "description": "Wallet login endpoint. Rate-limited to 5 requests per minute per client IP to mitigate brute-force attempts.",
        "security": [],
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "address",
                  "challenge",
                  "signature"
                ],
                "properties": {
                  "address": {
                    "type": "string",
                    "description": "Stellar public key.",
                    "example": "GABC1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ234567890ABCDEFG"
                  },
                  "challenge": {
                    "type": "string",
                    "description": "The nonce returned by GET /api/auth/wallet.",
                    "example": "streampay_auth_1716000000000_x7k2m"
                  },
                  "signature": {
                    "type": "string",
                    "description": "Base64-encoded Ed25519 signature of the challenge.",
                    "example": "dGVzdHNpZ25hdHVyZQ=="
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletToken"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "429": {
            "description": "IP rate limit exceeded for wallet login (5 req/min per IP).",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the client may retry.",
                "schema": {
                  "type": "integer"
                }
              },
              "x-request-id": {
                "description": "Correlation ID for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "error": {
                    "code": "rate_limit_exceeded",
                    "message": "Too many requests. Please try again later.",
                    "request_id": "req_01HZ9ABCDEF"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/api/debug/kms-sign": {
      "post": {
        "operationId": "debugKmsSign",
        "summary": "Sign a payload via KMS (debug only)",
        "description": "**Non-production only.** Returns a base64-encoded signature for the given payload. Returns 403 in production.",
        "security": [],
        "tags": [
          "Debug"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "payload"
                ],
                "properties": {
                  "payload": {
                    "type": "string",
                    "description": "Base64-encoded data to sign.",
                    "example": "aGVsbG8="
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signature produced.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "signature"
                  ],
                  "properties": {
                    "signature": {
                      "type": "string",
                      "description": "Base64-encoded signature.",
                      "example": "c2lnbmVkOmFHVnNiRzg5"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
...
}